Move unit tests into own files
